backend/requirements.txt
fastapi
uvicorn
pydantic
neo4j
supabase
python-jose[cryptography]
python-multipart
groq
langchain
PyMuPDF
python-dotenv
cors
